Fix a few size_t diag printf

+++ src/libnffile/nffile.c
@@ -596,7 +596,7 @@ static int ReadAppendix(nffile_t *nffile) {
 
     // seek back to currentPos
     off_t backPosition = lseek(nffile->fd, currentPos, SEEK_SET);
-    dbg_printf("Reset position to %lld -> %lld\n", currentPos, backPosition);
+    dbg_printf("Reset position to %lld -> %lld\n", (long long)currentPos, (long long)backPosition);
     if (backPosition < 0) {
         LogError("lseek() error in %s line %d: %s", __FILE__, __LINE__, strerror(errno));
         return 0;
